refactor: merge regular & animated label into Label.native.js
